requirements.txt
Django==1.8
psycopg2
Pillow
django-vanilla-views
django-forms-bootstrap
